+++ src/widgets/mdx/index.client.ts
@@ -1,0 +1,4 @@
+"use client";
+
+export { useMermaid } from "./helpers/plugins/mermaid/use-mermaid";
+export { useTwitter } from "./helpers/plugins/twitter/use-twitter";
