+++ src/widgets/mdx/index.ts
@@ -1,9 +1,4 @@
-export { exampleMDXComponent } from "./helpers/mocks/fixture-mdxcomponent";
-export { markupMermaid } from "./helpers/plugins/mermaid/mermaid-plugin";
-export { useMermaid } from "./helpers/plugins/mermaid/use-mermaid";
-export { useTwitter } from "./helpers/plugins/twitter/use-twitter";
 export { resolveMDXContent } from "./models/resolver";
-export { serializeMDXContent } from "./models/serializer";
 export type {
   MDXComponent,
   MDXCustomTextPlugin,
