+++ src/widgets/writings/models/writing-reader/read-writing.ts
@@ -5,7 +5,7 @@ import type {
   SerializedWriting,
   WritingHead,
 } from "../../../../entities/writing";
-import { markupMermaid, serializeMDXContent } from "../../../mdx";
+import { markupMermaid, serializeMDXContent } from "../../../mdx/index.server";
 import { mdxPlugins } from "../../ui/writing-detail/mdx/plugins";
 
 import { resolveWritingHead } from "./head-resolver";
